Break code out into separate modules
- Working towards pip install
- Add a setup.py script
